fix/refactor:ルームの遷移、表示をダイナミックルーティングに変更
